Updated ComputeEngineIntegrationTest
Fixed the test code so that the dimensions of the matrix pass the assertEquals tests for the rows having a length of 1 and the columns having a length of 10 for the test matrix.  Also fixed the assertEquals for an element of the matrix, since the elements of the matrix are randomly generated integers between 0-100 I changed the test code from assertEquals to assertTrue to test that the randomly generated integer is in the bounds of 0 and 100.